requirements.txt
discord.py
discord-pretty-help
PyNaCl
googlesearch-python
aiofiles
aiohttp
pymongo
motor
python-dotenv
openai